# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Investment Research Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ResearchInput {
  /**
     * Company name to research
     * @minLength 1
     */
  company: string;
}

export interface ResearchSession {
  id: number;
  company: string;
  /** pending | running | completed | failed */
  status: string;
  /**
     * INVEST | PASS | null
     * @nullable
     */
  verdict?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  reasoning?: string | null;
  /**
     * JSON array of bullet points supporting invest
     * @nullable
     */
  bulletsFor?: string | null;
  /**
     * JSON array of bullet points supporting pass
     * @nullable
     */
  bulletsAgainst?: string | null;
  /**
     * JSON blob of key financial metrics
     * @nullable
     */
  financialData?: string | null;
  /**
     * JSON array of recent news items
     * @nullable
     */
  newsHeadlines?: string | null;
  /**
     * 0-100 confidence in the verdict
     * @nullable
     */
  confidenceScore?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface ResearchStats {
  total: number;
  completed: number;
  investCount: number;
  passCount: number;
  /** @nullable */
  avgConfidence: number | null;
}