# lib/api-zod/src/generated/types/researchStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Investment Research Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchStats {
  total: number;
  completed: number;
  investCount: number;
  passCount: number;
  /** @nullable */
  avgConfidence: number | null;
}